LayoutTests/imported/w3c/web-platform-tests/css/cssom-view/image-x-y-zoom-expected.txt




FAIL x and y properties for image elements with css zoom assert_equals: image with css zoom in a no zoom container x expected 4 but got 8
